requirements.txt
streamlit
supabase
numpy
python-dotenv